third_party/blink/web_tests/external/wpt/html/semantics/popovers/popover-focus-2-expected.txt

This is a testharness.js-based test.
[FAIL] Popover focus navigation with imperative invocation
  assert_equals: Focus should move from invoker into the open popover expected Element node <button id="inside_popover1" tabindex="0">Inside1</button> but got Element node <button id="button3" tabindex="0">Button3</button>
[FAIL] Circular reference tab navigation with imperative invocation
  assert_equals: circular reference: Step 2 expected Element node <button id="circular1" autofocus="" popovertargetaction="... but got Element node <button id="circular4" tabindex="0">after</button>
[FAIL] Popover focus only returns to invoker when focus is within the popover with imperative invocation
  assert_equals: next up is the popover expected Element node <button tabindex="0">focusable element</button> but got Element node <span tabindex="0">Other focusable element</span>
[FAIL] Cases where the next focus candidate isn't in the direct parent scope with imperative invocation
  assert_equals: next up is the popover expected Element node <button tabindex="0">focusable element</button> but got Element node <body><div id="fixup">\n  <button id="button1" tabindex="0...
Harness: the test ran to completion.
